Extract query updates
